// src/app/models/person.ts
export interface Person {
  _id?: string;
  rol?: string;
  level_academy?: string;
  // specialty?: string,
  title: string;
  type_dni?: string;
  dni?: string;
  names: string;
  last_names: string;
  phone?: string;
  email?: string;
  password?: string;
  status?: boolean;
  congress_id: string;
}